fix: ENABLE_VOICE=true compile errors
- menu.c: guard MENU_SCR reference with #ifndef ENABLE_FEAT_F4HWN
  (MENU_SCR enum only exists when F4HWN is disabled)
- audio.c: rewrite AUDIO_PlaySingleVoice and AUDIO_PlayQueuedVoice
  to use PY32 SPI-flash voice mechanism (LoadVoiceClip/VOICE_Start)
  instead of removed DP32 VoiceClipLengthChinese/English arrays and
  AUDIO_PlayVoice function
- bitmaps.h/c: add missing BITMAP_VoicePrompt[7] under ENABLE_VOICE
  (referenced in ui/status.c but never defined in bafang port)
- driver/voice.c: cast DAC_Buf pointer to uint32_t for
  LL_DMA_ConfigAddresses (pointer/integer mismatch)

+++ App/driver/voice.c
@@ -118,7 +118,7 @@ void VOICE_Start()
         memset(DAC_Buf + VOICE_BUF_SIZE, 0, VOICE_BUF_SIZE);
     }
 
-    LL_DMA_ConfigAddresses(DMA1, DMA_CHANNEL, DAC_Buf,                                                         //
+    LL_DMA_ConfigAddresses(DMA1, DMA_CHANNEL, (uint32_t)(uintptr_t)DAC_Buf,                                     //
                            LL_DAC_DMA_GetRegAddr(DAC1, DAC_CHANNEL, LL_DAC_DMA_REG_DATA_12BITS_RIGHT_ALIGNED), //
                            LL_DMA_DIRECTION_MEMORY_TO_PERIPH                                                   //
     );
